Content Updated in ViewSignInPage
