+++ lib/ui/root_page/root_page.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:training_app/ui/sign_in_page/sign_in_page.dart';
+import '../sign_in_page/sign_in_page_provider.dart';
 
 class DearDiaryApp extends StatelessWidget {
   const DearDiaryApp({super.key});
@@ -14,7 +14,7 @@ class DearDiaryApp extends StatelessWidget {
         ),
         home: const Scaffold(
           backgroundColor: Colors.blue,
-          body: SignInPage(),
+          body: SignInPageProvider(),
         ));
   }
 }
